QA fixes, fix migration files

+++ migrations/20180621213451_add_transaction_data_to_treasures.up.sql
@@ -5,16 +5,16 @@ ALTER TABLE `treasures`
   ADD COLUMN IF NOT EXISTS `prl_tx_hash` VARCHAR (255) DEFAULT NULL;
 
 ALTER TABLE `treasures`
-  ADD COLUMN IF NOT EXISTS `prl_tx_nonce` decimal (28, 18) DEFAULT NULL;
+  ADD COLUMN IF NOT EXISTS `prl_tx_nonce` bigint (20) DEFAULT NULL;
 
 ALTER TABLE `treasures`
   ADD COLUMN IF NOT EXISTS `gas_tx_hash` VARCHAR (255) DEFAULT NULL;
 
 ALTER TABLE `treasures`
-  ADD COLUMN IF NOT EXISTS `gas_tx_nonce` decimal (28, 18) DEFAULT NULL;
+  ADD COLUMN IF NOT EXISTS `gas_tx_nonce` bigint (20) DEFAULT NULL;
 
 ALTER TABLE `treasures`
   ADD COLUMN IF NOT EXISTS `bury_tx_hash` VARCHAR (255) DEFAULT NULL;
 
 ALTER TABLE `treasures`
-  ADD COLUMN IF NOT EXISTS `bury_tx_nonce` decimal (28, 18) DEFAULT NULL;
+  ADD COLUMN IF NOT EXISTS `bury_tx_nonce` bigint (20) DEFAULT NULL;
