+++ migrations/20180621213452_add_transaction_data_to_completed_uploads.up.sql
@@ -2,10 +2,10 @@ ALTER TABLE `completed_uploads`
   ADD COLUMN IF NOT EXISTS `prl_tx_hash` VARCHAR (255) DEFAULT NULL;
 
 ALTER TABLE `completed_uploads`
-  ADD COLUMN IF NOT EXISTS `prl_tx_nonce` decimal (28, 18) DEFAULT NULL;
+  ADD COLUMN IF NOT EXISTS `prl_tx_nonce` bigint (20) DEFAULT NULL;
 
 ALTER TABLE `completed_uploads`
   ADD COLUMN IF NOT EXISTS `gas_tx_hash` VARCHAR (255) DEFAULT NULL;
 
 ALTER TABLE `completed_uploads`
-  ADD COLUMN IF NOT EXISTS `gas_tx_nonce` decimal (28, 18) DEFAULT NULL;
+  ADD COLUMN IF NOT EXISTS `gas_tx_nonce` bigint (20) DEFAULT NULL;
